Remove race condition between reading updates to config-observability and updating stackdriver exporter
