+++ metrics/config.go
@@ -27,7 +27,6 @@ import (
 	"time"
 
 	"go.uber.org/zap"
-	corev1 "k8s.io/api/core/v1"
 )
 
 const (
@@ -59,34 +58,6 @@ const (
 	maxPrometheusPort     = 65535
 	minPrometheusPort     = 1024
 )
-
-// ExporterOptions contains options for configuring the exporter.
-type ExporterOptions struct {
-	// Domain is the metrics domain. e.g. "knative.dev". Must be present.
-	//
-	// Stackdriver uses the following format to construct full metric name:
-	//    <domain>/<component>/<metric name from View>
-	// Prometheus uses the following format to construct full metric name:
-	//    <component>_<metric name from View>
-	// Domain is actually not used if metrics backend is Prometheus.
-	Domain string
-
-	// Component is the name of the component that emits the metrics. e.g.
-	// "activator", "queue_proxy". Should only contains alphabets and underscore.
-	// Must be present.
-	Component string
-
-	// PrometheusPort is the port to expose metrics if metrics backend is Prometheus.
-	// It should be between maxPrometheusPort and maxPrometheusPort. 0 value means
-	// using the default 9090 value. If is ignored if metrics backend is not
-	// Prometheus.
-	PrometheusPort int
-
-	// ConfigMap is the data from config map config-observability. Must be present.
-	// See https://github.com/knative/serving/blob/master/config/config-observability.yaml
-	// for details.
-	ConfigMap map[string]string
-}
 
 type metricsConfig struct {
 	// The metrics domain. e.g. "serving.knative.dev" or "build.knative.dev".
@@ -131,7 +102,7 @@ type metricsConfig struct {
 	stackdriverCustomMetricTypePrefix string
 }
 
-func getMetricsConfig(ops ExporterOptions, logger *zap.SugaredLogger) (*metricsConfig, error) {
+func createMetricsConfig(ops ExporterOptions, logger *zap.SugaredLogger) (*metricsConfig, error) {
 	var mc metricsConfig
 
 	if ops.Domain == "" {
@@ -219,62 +190,6 @@ func getMetricsConfig(ops ExporterOptions, logger *zap.SugaredLogger) (*metricsC
 	}
 
 	return &mc, nil
-}
-
-// UpdateExporterFromConfigMap returns a helper func that can be used to update the exporter
-// when a config map is updated.
-func UpdateExporterFromConfigMap(component string, logger *zap.SugaredLogger) func(configMap *corev1.ConfigMap) {
-	domain := Domain()
-	return func(configMap *corev1.ConfigMap) {
-		UpdateExporter(ExporterOptions{
-			Domain:    domain,
-			Component: component,
-			ConfigMap: configMap.Data,
-		}, logger)
-	}
-}
-
-// UpdateExporter updates the exporter based on the given ExporterOptions.
-func UpdateExporter(ops ExporterOptions, logger *zap.SugaredLogger) error {
-	newConfig, err := getMetricsConfig(ops, logger)
-	if err != nil {
-		if ce := getCurMetricsExporter(); ce == nil {
-			// Fail the process if there doesn't exist an exporter.
-			logger.Errorw("Failed to get a valid metrics config", zap.Error(err))
-		} else {
-			logger.Errorw("Failed to get a valid metrics config; Skip updating the metrics exporter", zap.Error(err))
-		}
-		return err
-	}
-
-	if isNewExporterRequired(newConfig) {
-		logger.Info("Flushing the existing exporter before setting up the new exporter.")
-		FlushExporter()
-		e, err := newMetricsExporter(newConfig, logger)
-		if err != nil {
-			logger.Errorf("Failed to update a new metrics exporter based on metric config %v. error: %v", newConfig, err)
-			return err
-		}
-		existingConfig := getCurMetricsConfig()
-		setCurMetricsExporter(e)
-		logger.Infof("Successfully updated the metrics exporter; old config: %v; new config %v", existingConfig, newConfig)
-	}
-
-	setCurMetricsConfig(newConfig)
-	return nil
-}
-
-// isNewExporterRequired compares the non-nil newConfig against curMetricsConfig. When backend changes,
-// or stackdriver project ID changes for stackdriver backend, we need to update the metrics exporter.
-func isNewExporterRequired(newConfig *metricsConfig) bool {
-	cc := getCurMetricsConfig()
-	if cc == nil || newConfig.backendDestination != cc.backendDestination {
-		return true
-	} else if newConfig.backendDestination == Stackdriver && newConfig.stackdriverProjectID != cc.stackdriverProjectID {
-		return true
-	}
-
-	return false
 }
 
 // ConfigMapName gets the name of the metrics ConfigMap
